// windows-edition/teldrive-ui/src/lib/api.d.ts
export interface components {
    schemas: {
// ...
        FileList: {
            /** @description Array of file entries in the current page */
            items: components["schemas"]["File"][];
            /** @description Pagination metadata and total count information */
            meta: components["schemas"]["Meta"];
        };
// ...
        SessionCreate: {
// ...
            session: string;
        };
// ...
    };
// ...
}